+++ source/Irrlicht/CParticleSystemSceneNode.cpp
@@ -524,12 +524,12 @@ void CParticleSystemSceneNode::reallocateBuffers()
 
 		for (i=oldIdxSize; i<Buffer->Indices.size(); i+=6)
 		{
-			Buffer->Indices[0+i] = 0+oldvertices;
-			Buffer->Indices[1+i] = 2+oldvertices;
-			Buffer->Indices[2+i] = 1+oldvertices;
-			Buffer->Indices[3+i] = 0+oldvertices;
-			Buffer->Indices[4+i] = 3+oldvertices;
-			Buffer->Indices[5+i] = 2+oldvertices;
+			Buffer->Indices[0+i] = (u16)0+oldvertices;
+			Buffer->Indices[1+i] = (u16)2+oldvertices;
+			Buffer->Indices[2+i] = (u16)1+oldvertices;
+			Buffer->Indices[3+i] = (u16)0+oldvertices;
+			Buffer->Indices[4+i] = (u16)3+oldvertices;
+			Buffer->Indices[5+i] = (u16)2+oldvertices;
 			oldvertices += 4;
 		}
 	}
