+++ source/Irrlicht/CVolumeLightSceneNode.cpp
@@ -53,7 +53,7 @@ void CVolumeLightSceneNode::addToBuffer(const video::S3DVertex& v)
 	const bool alreadyIn = (tnidx != -1);
 	u16 nidx = (u16)tnidx;
 	if (!alreadyIn) {
-		nidx = Buffer->Vertices.size();
+		nidx = (u16)Buffer->Vertices.size();
 		Buffer->Indices.push_back(nidx);
 		Buffer->Vertices.push_back(v);
 	} else
